NEXUS-4744: factoring out the background task spawning code
